+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.4.11
 
+* Fix a bug where it was possible to reject recourse, even though it was already rejected
+
 # 0.4.10
 
 * Recoursee in a request to recourse does not have to be in the contact book anymore