+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.4.11
 
 * Fix a bug where it was possible to reject recourse, even though it was already rejected
+* Fixed an issue where it could happen that identity and company contacts weren't propagated to Nostr, leading to block propagation inconsistencies
 
 # 0.4.10